+++ src/Meta/Render/Meshes/VoxelMesh.interface.ts
@@ -18,7 +18,7 @@ export type VoxelMeshInterface = {
 export type MeshSetData = {
  positionArray: Float32Array;
  normalsArray: Float32Array;
- indiciesArray: Int32Array;
+ indiciesArray: Uint16Array;
  faceDataArray: Float32Array;
  AOColorsArray: Float32Array;
  lightColorsArray: Float32Array;
